Move Login & OAuth to shared/
