+++ DataManagement/src/components/Datastream/DatastreamSourceLinker.vue
@@ -51,7 +51,7 @@
 <script setup lang="ts">
 import { onMounted, ref } from 'vue'
 import { api } from '@shared/services/api'
-import { rules } from '@/utils/rules'
+import { rules } from '@shared/utils/rules'
 import { DataSource, Datastream } from '@shared/types'
 import { useFormLogic } from '@/composables/useFormLogic'
 import { Snackbar } from '@shared/utils/notifications'
