+++ DataManagement/src/components/HydroShare/HydroShareFormCard.vue
@@ -197,7 +197,7 @@ import { onMounted, ref } from 'vue'
 import { Datastream, PostHydroShareArchive, Frequency } from '@shared/types'
 import { api } from '@shared/services/api'
 import { VForm } from 'vuetify/components'
-import { hydroShareUrl, rules } from '@/utils/rules'
+import { hydroShareUrl, rules } from '@shared/utils/rules'
 import { useFormLogic } from '@/composables/useFormLogic'
 import HydroShareDeleteCard from '@/components/HydroShare/HydroShareDeleteCard.vue'
 import { useHydroShareStore } from '@/store/hydroShare'
