+++ DataManagement/src/components/Metadata/UnitFormCard.vue
@@ -55,7 +55,7 @@
 </template>
 
 <script setup lang="ts">
-import { rules } from '@/utils/rules'
+import { rules } from '@shared/utils/rules'
 import { api } from '@shared/services/api'
 import { VForm } from 'vuetify/components'
 import { useFormLogic } from '@/composables/useFormLogic'
