+++ DataManagement/src/components/Site/SiteForm.vue
@@ -167,7 +167,7 @@ import { useThingStore } from '@/store/thing'
 import { Thing } from '@shared/types'
 import { siteTypes } from '@/vocabularies'
 import { VForm } from 'vuetify/components'
-import { rules } from '@/utils/rules'
+import { rules } from '@shared/utils/rules'
 import { storeToRefs } from 'pinia'
 import { api } from '@shared/services/api'
 import SitePhotoManager from '@/components/Site/SitePhotoManager.vue'
