+++ DataManagement/src/pages/DatastreamForm.vue
@@ -318,7 +318,7 @@ import SensorFormCard from '@/components/Metadata/SensorFormCard.vue'
 import ObservedPropertyFormCard from '@/components/Metadata/ObservedPropertyFormCard.vue'
 import UnitFormCard from '@/components/Metadata/UnitFormCard.vue'
 import ProcessingLevelFormCard from '@/components/Metadata/ProcessingLevelFormCard.vue'
-import { rules } from '@/utils/rules'
+import { rules } from '@shared/utils/rules'
 import { mediumTypes, aggregationTypes, statusTypes } from '@/vocabularies'
 import { useMetadata } from '@/composables/useMetadata'
 import { Thing, Unit } from '@shared/types'
