+++ shared/src/router/routes.ts
@@ -1,3 +1,13 @@
 import { RouteRecordRaw } from 'vue-router'
 
-export const routes: RouteRecordRaw[] = []
+export const routes: RouteRecordRaw[] = [
+  {
+    path: '/login',
+    name: 'Login',
+    component: () => import('@shared/pages/account/Login.vue'),
+    meta: {
+      title: 'Login',
+      hasLoggedOutGuard: true,
+    },
+  },
+]
