+++ shared/src/utils/rules.ts
